+++ hurricane/include/hurricane/renderer/renderer.h
@@ -1,9 +1,9 @@
 #ifndef HC_RENDERER_H
 #define HC_RENDERER_H
 
-double calculate_depth(double x0, double y0, double z0, double x1, double y1,
-                       double z1, double x2, double y2, double z2, double px,
-                       double py);
+float calculate_depth(float x0, float y0, float z0, float x1, float y1,
+                       float z1, float x2, float y2, float z2, float px,
+                       float py);
 
 typedef struct hc_renderer_config {
   int width;
@@ -14,11 +14,11 @@ typedef struct hc_renderer_config {
 typedef struct hc_renderer {
   void (*init)(hc_renderer_config);
   void (*pre_frame)(void);
-  void (*triangle)(int, int, double, int, int, double, int, int, double,
+  void (*triangle)(int, int, float, int, int, float, int, int, float,
                    unsigned char, unsigned char, unsigned char);
   void (*frame)(void);
   void (*finish)(void);
-  double (**internal_depth_buf);
+  float (**internal_depth_buf);
 } hc_renderer;
 
 #endif
